requirements.txt
discord.py
pyke-lol
python-dotenv
